Fixed: warnings with printf(%d, size_t) Fixed: All copyrights are now correct up to 2007 Fixed: Build system now works with VPATHs Changed: protocol_example is now ignored by default
This commit was SVN r15627.

+++ ompi/mca/pml/v/protocol_example/vprotocol_example_proc.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2004-2005 The Trustees of the University of Tennessee.
+ * Copyright (c) 2004-2007 The Trustees of the University of Tennessee.
  *                         All rights reserved.
  * $COPYRIGHT$
  *
@@ -14,12 +14,12 @@
 
 int mca_vprotocol_example_add_procs(struct ompi_proc_t **procs, size_t nprocs)
 {
-  V_OUTPUT_VERBOSE(30, "adding %d procs", nprocs);
+  V_OUTPUT_VERBOSE(30, "adding %ld procs", (long) nprocs);
   return mca_pml_v.host_pml.pml_add_procs(procs, nprocs);
 }
 
 int mca_vprotocol_example_del_procs(struct ompi_proc_t **procs, size_t nprocs)
 {
-  V_OUTPUT_VERBOSE(30, "removing %d procs", nprocs);
+  V_OUTPUT_VERBOSE(30, "removing %ld procs", (long) nprocs);
   return mca_pml_v.host_pml.pml_del_procs(procs, nprocs);
 }
